ingestion-api/requirements.txt
fastapi
uvicorn
msgpack
pydantic
pytest
psycopg2-binary
